+++ weblate/formats/external.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 import os
 from io import BytesIO, StringIO
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
 from zipfile import BadZipFile
 
 from django.utils.translation import gettext_lazy
@@ -137,6 +137,7 @@ class XlsxFormat(CSVUtf8Format):
         language: str,
         base: str,
         callback: Callable | None = None,
+        file_format_params: dict[str, Any] | None = None,  # noqa: ARG003
     ) -> None:
         """Handle creation of new translation file."""
         if not base: